fix: 一些同步和资源访问追踪的问题。

+++ mi/renderer/mi_renderer_view.cpp
@@ -157,7 +157,7 @@ void BatchedUploadContext::Fire(RenderGraphBuilder &builder) {
                 if (buffer.offset + buffer.size < e.dst_buffer.offset + e.data.size()) {
                     buffer = e.dst_buffer;
                     buffer.size = e.data.size();
-                    furthest_index = i;
+                    furthest_index = (int)i;
                 }
             }
         }
